docs/archive/requirements.txt
guzzle_sphinx_theme
git+git://github.com/exponent/sphinx.git@master
pygments-lexer-babylon
sphinx-autobuild
pygments-style-exponent
